analyzer: Move dependency handlers to the `utils` package
The top-level `managers` package should only contain the core package
manager implementations themselves for a better overview.
